+++ src/scranpy/lib/bindings.cpp
@@ -159,7 +159,7 @@ void run_umap(void*, int32_t);
 
 void scale_by_neighbors(int32_t, const uintptr_t*, int32_t, double*, int32_t);
 
-void score_feature_set(void*, const int32_t*, uint8_t, const int32_t*, double*, double*, uint8_t, int32_t);
+void score_feature_set(void*, const uint8_t*, uint8_t, const int32_t*, double*, double*, uint8_t, int32_t);
 
 void score_markers(const void*, int32_t, const int32_t*, int32_t, const int32_t*, uint8_t, double, uintptr_t*, uintptr_t*, uintptr_t*, uintptr_t*, uintptr_t*, uintptr_t*, int32_t);
 
@@ -1111,7 +1111,7 @@ PYAPI void py_scale_by_neighbors(int32_t nembeddings, const uintptr_t* indices, 
     }
 }
 
-PYAPI void py_score_feature_set(void* mat, const int32_t* features, uint8_t use_block, const int32_t* block, double* output_scores, double* output_weights, uint8_t scale, int32_t nthreads, int32_t* errcode, char** errmsg) {
+PYAPI void py_score_feature_set(void* mat, const uint8_t* features, uint8_t use_block, const int32_t* block, double* output_scores, double* output_weights, uint8_t scale, int32_t nthreads, int32_t* errcode, char** errmsg) {
     try {
         score_feature_set(mat, features, use_block, block, output_scores, output_weights, scale, nthreads);
     } catch(std::exception& e) {
